refactor: SessionDetail・MetricsPage のコンポーネント分離 (#148)
