+++ frontend/src/components/session/index.ts
@@ -1,0 +1,5 @@
+export { ToolInputView } from "./ToolInputView";
+export { ToolCallView } from "./ToolCallView";
+export { StreamDisplayItemView } from "./StreamDisplayItemView";
+export { StreamOutputView } from "./StreamOutputView";
+export { DiffDropdown } from "./DiffDropdown";
